// src/resgatpet/dto/formulario/listaFormulario.dto.d.ts
export declare class ListaFormularioDTO {
    readonly id: string;
    readonly fotoPet: string;
    readonly endereco: string;
    readonly cidade: string;
    readonly raca: string;
    readonly sexo: string;
    readonly cor: string;
    readonly saude: string;
    readonly acessorio: string;
    readonly usuario: string;
    constructor(id: string, fotoPet: string, endereco: string, cidade: string, raca: string, sexo: string, cor: string, saude: string, acessorio: string, usuario: string);
}