// src/resgatpet/formulario/formulario.entity.d.ts
export declare class FormularioEntity {
    id: string;
    fotoPet: string;
    endereco: string;
    cidade: string;
    raca: string;
    sexo: string;
    cor: string;
    saude: string;
    acessorio: string;
    usuario: string;
    constructor(id: string, fotoPet: string, endereco: string, cidade: string, raca: string, sexo: string, cor: string, saude: string, acessorio: string, usuario: string);
}